sql: Fixed parse_tags function

+++ sql/migrate-5.59.sql
@@ -10,7 +10,7 @@ CREATE FUNCTION iris.parse_tags(notes TEXT) RETURNS SETOF TEXT AS
     $parse_tags$
 BEGIN
     RETURN QUERY SELECT tag[1] FROM (
-        SELECT regexp_matches(notes, '#([A-Za-z0-9]+)', 'g') AS tag
+        SELECT regexp_matches(notes, '(#[A-Za-z0-9]+)', 'g') AS tag
     ) AS tags;
 END;
 $parse_tags$ LANGUAGE plpgsql STABLE;
@@ -18,13 +18,15 @@ $parse_tags$ LANGUAGE plpgsql STABLE;
 CREATE OR REPLACE FUNCTION iris.camera_hashtag() RETURNS TRIGGER AS
     $camera_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'camera' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'camera', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'camera' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'camera', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -33,13 +35,15 @@ $camera_hashtag$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION iris.beacon_hashtag() RETURNS TRIGGER AS
     $beacon_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'beacon' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'beacon', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'beacon' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'beacon', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -48,13 +52,15 @@ $beacon_hashtag$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION iris.gps_hashtag() RETURNS TRIGGER AS
     $gps_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'gps' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'gps', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'gps' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'gps', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -63,13 +69,15 @@ $gps_hashtag$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION iris.dms_hashtag() RETURNS TRIGGER AS
     $dms_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'dms' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'dms', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'dms' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'dms', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -78,13 +86,15 @@ $dms_hashtag$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION iris.gate_arm_array_hashtag() RETURNS TRIGGER AS
     $gate_arm_array_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'gate_arm_array' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'gate_arm_array', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'gate_arm_array' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'gate_arm_array', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -93,13 +103,15 @@ $gate_arm_array_hashtag$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION iris.ramp_meter_hashtag() RETURNS TRIGGER AS
     $ramp_meter_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'ramp_meter' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'ramp_meter', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'ramp_meter' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'ramp_meter', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -108,13 +120,15 @@ $ramp_meter_hashtag$ LANGUAGE plpgsql;
 CREATE OR REPLACE FUNCTION iris.weather_sensor_hashtag() RETURNS TRIGGER AS
     $weather_sensor_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'weather_sensor' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'weather_sensor', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'weather_sensor' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'weather_sensor', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
