+++ sql/tms-template.sql
@@ -499,7 +499,7 @@ CREATE FUNCTION iris.parse_tags(notes TEXT) RETURNS SETOF TEXT AS
     $parse_tags$
 BEGIN
     RETURN QUERY SELECT tag[1] FROM (
-        SELECT regexp_matches(notes, '#([A-Za-z0-9]+)', 'g') AS tag
+        SELECT regexp_matches(notes, '(#[A-Za-z0-9]+)', 'g') AS tag
     ) AS tags;
 END;
 $parse_tags$ LANGUAGE plpgsql STABLE;
@@ -1678,13 +1678,15 @@ ALTER TABLE iris._camera ADD CONSTRAINT _camera_fkey
 CREATE FUNCTION iris.camera_hashtag() RETURNS TRIGGER AS
     $camera_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'camera' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'camera', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'camera' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'camera', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -2165,13 +2167,15 @@ ALTER TABLE iris._beacon ADD CONSTRAINT _beacon_fkey
 CREATE FUNCTION iris.beacon_hashtag() RETURNS TRIGGER AS
     $beacon_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'beacon' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'beacon', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'beacon' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'beacon', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -2567,13 +2571,15 @@ ALTER TABLE iris._gps ADD CONSTRAINT _gps_fkey
 CREATE FUNCTION iris.gps_hashtag() RETURNS TRIGGER AS
     $gps_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'gps' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'gps', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'gps' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'gps', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -3204,13 +3210,15 @@ GRANT SELECT ON dms_message_view TO PUBLIC;
 CREATE FUNCTION iris.dms_hashtag() RETURNS TRIGGER AS
     $dms_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'dms' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'dms', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'dms' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'dms', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -3414,13 +3422,15 @@ ALTER TABLE iris._gate_arm_array ADD CONSTRAINT _gate_arm_array_fkey
 CREATE FUNCTION iris.gate_arm_array_hashtag() RETURNS TRIGGER AS
     $gate_arm_array_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'gate_arm_array' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'gate_arm_array', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'gate_arm_array' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'gate_arm_array', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -4801,13 +4811,15 @@ ALTER TABLE iris._ramp_meter ADD CONSTRAINT _ramp_meter_fkey
 CREATE FUNCTION iris.ramp_meter_hashtag() RETURNS TRIGGER AS
     $ramp_meter_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'ramp_meter' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'ramp_meter', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'ramp_meter' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'ramp_meter', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
@@ -5471,13 +5483,15 @@ ALTER TABLE iris._weather_sensor ADD CONSTRAINT _weather_sensor_fkey
 CREATE FUNCTION iris.weather_sensor_hashtag() RETURNS TRIGGER AS
     $weather_sensor_hashtag$
 BEGIN
-    IF (TG_OP != 'INSERT') THEN
-        DELETE FROM iris.hashtag
-        WHERE resource_n = 'weather_sensor' AND name = OLD.name;
-    END IF;
-    IF (TG_OP != 'DELETE') THEN
-        INSERT INTO iris.hashtag (resource_n, name, hashtag)
-        SELECT 'weather_sensor', NEW.name, iris.parse_tags(NEW.notes);
+    IF (NEW.notes IS DISTINCT FROM OLD.notes) THEN
+        IF (TG_OP != 'INSERT') THEN
+            DELETE FROM iris.hashtag
+            WHERE resource_n = 'weather_sensor' AND name = OLD.name;
+        END IF;
+        IF (TG_OP != 'DELETE') THEN
+            INSERT INTO iris.hashtag (resource_n, name, hashtag)
+            SELECT 'weather_sensor', NEW.name, iris.parse_tags(NEW.notes);
+        END IF;
     END IF;
     RETURN NULL; -- AFTER trigger return is ignored
 END;
